refactor: modularize imports and update configuration references
