+++ module/data/item/spell.mjs
@@ -4,9 +4,9 @@ import LearnSpellPrompt from "../../applications/advancement/learn-spell.mjs";
 import TargetTemplate from "./templates/targeting.mjs";
 import { AreaMetricData, DurationMetricData, MetricData, RangeMetricData } from "../common/metrics.mjs";
 import ItemDataModel from "../abstract/item-data-model.mjs";
-import { SelectExtraThreadsPrompt } from "../../applications/workflow/_module.mjs";
+import SelectExtraThreadsPrompt from "../../applications/workflow/select-extra-threads-prompt.mjs";
 import ThreadWeavingRollOptions from "../roll/weaving.mjs";
-import { RollPrompt } from "../../applications/global/_module.mjs";
+import RollPrompt from "../../applications/global/roll-prompt.mjs";
 import SpellcastingRollOptions from "../roll/spellcasting.mjs";
 import RollProcessor from "../../services/roll-processor.mjs";
 import SpellEffectRollOptions from "../roll/spelleffect.mjs";
