+++ module/documents/actor.mjs
@@ -11,7 +11,9 @@ import PromptFactory from "../applications/global/prompt-factory.mjs";
 import ClassTemplate from "../data/item/templates/class.mjs";
 import MigrationManager from "../services/migrations/migration-manager.mjs";
 import AttackWorkflow from "../workflows/workflow/attack-workflow.mjs";
-import { AttributeWorkflow, AttuneMatrixWorkflow, KnockdownWorkflow } from "../workflows/workflow/_module.mjs";
+import AttributeWorkflow from "../workflows/workflow/attribute-workflow.mjs";
+import AttuneMatrixWorkflow from "../workflows/workflow/attune-matrix-workflow.mjs";
+import KnockdownWorkflow from "../workflows/workflow/knockdown-workflow.mjs";
 import { getDefaultEdid, getSetting } from "../settings.mjs";
 import RollProcessor from "../services/roll-processor.mjs";
 import RecoveryWorkflow from "../workflows/workflow/recovery-workflow.mjs";
